Create github-pages.yml (#109)
Builds site and commits static files to gh-pages branch on every commit to master branch.

Github pages only allows a small set of plugins. We recently added jekyll-tabs in #95, and will probably add more in the future. If this action works well, will update the site settings to serve what is in gh-pages branch.