Decision doc (#114)
In discussion on https://reviewable.io/reviews/datacommonsorg/docsite/97, it became clear that some sort of standard was needed for when to describe (for example) objects as places or Place objects. After thinking about, it seemed to me that having a separate decision document for contributors to refer to might be a good idea--for example, if we note decisions in the readme, it might become cluttered.